Polishing Trigger Events and Organizing Assets
Due to the previous method of collecting a collectible the player would lose momentum if they collided with it, the method was overhauled to have OnTriggerEnter() handle the colision on individual assets so the player continues to move unimpeded.

Assets required organization for level design, collectibles now have their own sub folder.